product-service/requirements.txt
Django==5.0.*
djangorestframework
psycopg[binary]
redis
django-redis
python-dotenv
django-filter
faker
djangorestframework-simplejwt[crypto]